Handle error in `command.handle`
Prior to this change, all errors that occured during the execution of
a command's `handle` method were simply left unhandled.

This change generally keeps that behaviour (with an added call to
logger) however does handle some special handling for the
`CommandVerificationError` which is a subset of the overall
`CommandError` used to indicate a user error. For these kinds of
errors we return a 500 to the user with the error message to help the
user understand their mistake.