Improve league table tests
Prior to this change, the tests checked how the exact string that was
put out by tabulate was formatted.

Ideally we still want to check the actual content output of the
command (that is the point of the test after all) but it would be nice
if the assertions in the test weren't so picky about the exact spacing
used by tabulate, amongst other things.

This change adds a function to strip away some of the formatting to
leave the table in a line by line, column by column 2D list for us to
assert against.