Add `build_tally_data_by_player` query function
This change adds a new query to the match tracker app which pulls
interesting data out of a series of sorted matches.

The data is returned as a dict keyed on the `User` instance where the
value is an instance of `MatchesTallyData` which is a dataclass
containing the info pulled from the list of match results.

`build_tally_data_by_player` assumes that the match results provided
to it are sorted by `played_at` datetime since it assumes that matches
laster in the list come after those earlier in the list.